Fix week 1 starting in december
If week 1 starts in december, set the year to next year, to
fetch correct schedule from Zermelo

Example:
In 2026 week 1 starts in december 2025. Week===1 is in december.
We need to fetch the liveschedule of week=202601 instead of 202501

Also (without this fix) klascal will jump to week 2 of 2025
(instead of 2026) when moving to the next week.